MotionBlock component and cleaned up Preview and removed BlockIterator
