--- src/modules/blocks/blocks.mod.scss
@@ -1,12 +1,0 @@
-@import '~mixins/flex';
-@import '~mixins/center';
-
-.wrapper {
-  @include flex;
-  @include center;
-
-  height: inherit;
-  width: inherit;
-
-  padding-top: 20px;
-}
